fix(lineage): repair SQLite autoincrement + run-view header URL
Two bugs surfaced in the Phase 15 live E2E replay (headful Firefox
walkthrough of the medallion notebook):

1. lineage_row_edges.id used BigInteger PK, which does not autoincrement
   in SQLite (only INTEGER PRIMARY KEY does). Symptom: every merge call
   stamped [lineage_edges_partial] IntegrityError on agent_run_operations
   and produced zero per-row edges. Fix: switch the model column and the
   matching Alembic migration body to Integer.

2. run_view.html header link "View lineage graph" pointed at
   /catalogs/{cat}/{schema}/{table} (no /schemas/ + /tables/ segments),
   producing 404 on click. Line 594 of the same template already had the
   correct pattern; line 87 was a copy-paste miss. Fix: align line 87
   with the rest of the codebase.

Both bugs were architectural blind spots that ruff/pyright/pydoclint
cannot catch (SQLite quirk + URL string template) and were caught only
by running the full notebook + clicking through the UI in a real
browser. Reinforces "live replay as gate" memo.

+++ pointlessql/alembic/versions/d4e5f6a7b8c9_lineage_row_edges.py
@@ -32,7 +32,7 @@ depends_on: str | Sequence[str] | None = None
 def upgrade() -> None:
     op.create_table(
         "lineage_row_edges",
-        sa.Column("id", sa.BigInteger(), autoincrement=True, nullable=False),
+        sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
         sa.Column("run_id", sa.String(length=36), nullable=False),
         sa.Column("op_id", sa.Integer(), nullable=False),
         sa.Column("source_table", sa.String(length=255), nullable=False),
